services/enroller/requirements.txt
# enroller Python dependencies
fastapi>=0.111.0
uvicorn>=0.30.0
asyncpg>=0.29.0
httpx>=0.27.0
scrapli[asyncssh]==2024.1.30
prometheus-fastapi-instrumentator==7.0.0
python-nmap==0.7.1
cryptography==42.0.8
apscheduler==3.10.4
